test-requirements.txt
# The order of packages is significant, because pip processes them in the order
# of appearance. Changing the order has an impact on the overall integration
# process, which may cause wedges in the gate later.
# Hacking already pins down pep8, pyflakes and flake8
hacking>=0.9.2,<0.10
# Testing Requirements
coverage>=3.6
django-nose
mock>=1.0
mox>=0.5.3
nodeenv>=0.9.4 # BSD License
nose
nose-exclude
nosexcover
openstack.nose_plugin>=0.7
nosehtmloutput>=0.0.3
selenium
xvfbwrapper>=0.1.3 #license: MIT
# Docs Requirements
sphinx>=1.1.2,!=1.2.0,<1.3
oslosphinx>=2.2.0  # Apache-2.0

http://tarballs.openstack.org/horizon/horizon-master.tar.gz#egg=horizon
